pyproject.toml
[project]
name = "perps-funding-arb"
version = "0.1.0"
description = "Cross-exchange perpetual futures funding rate arbitrage research"
requires-python = ">=3.11"
dependencies = [
    # Core
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    # Exchange APIs
    "ccxt>=4.0",
    "requests>=2.31",
    # Visualization
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotly>=5.15",
    # Notebooks
    "jupyter>=1.0",
    "jupyterlab>=4.0",
    # Utilities
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "pyarrow>=13.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
